+++ magicarp/error_codes.py
@@ -1,6 +1,0 @@
-SCHEMA_CODES = {
-}
-
-
-ATTR_CODES = {
-}