+++ magicarp/signals.py
@@ -1,0 +1,5 @@
+from flask import signals
+
+namespace = signals.Namespace()
+
+app_shutdown = namespace.signal('app_shutdown')