+++ HISTORY.rst
@@ -5,6 +5,9 @@ New Features:
 
  * input/output fields on schema that allow blanks will return None if None was
    provided (with possibly default to be set for such cases in the future)
+ * automated error codes for incomming data validation are possible to override,
+   change ERROR_CODEBOOK_ATTRIBUTES and ERROR_CODEBOOK_SCHEMA in your local
+   settings
 
 1.2.0 (2018-06-07)
 ~~~~~~~~~~~~~~~~~~